Update `dials.combine_experiments` (#2323)
Fix to allow combine experiments with the need for reflection tables.

Fixes #1741
Fixes #1658
