+++ newsfragments/2323.bugfix
@@ -1,0 +1,1 @@
+``dials.combine_experiments``: Fix to not require reflection data
