+++ src/dials/algorithms/indexing/ssx/processing.py
@@ -19,7 +19,7 @@ from dials.algorithms.indexing import DialsIndexError
 from dials.algorithms.indexing.indexer import Indexer
 from dials.algorithms.indexing.max_cell import find_max_cell
 from dials.array_family import flex
-from dials.command_line.combine_experiments import CombineWithReference
+from dials.util.combine_experiments import CombineWithReference
 
 RAD2DEG = 180 / math.pi
